requirements.txt
a2a-sdk==0.2.8
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Authlib==1.6.0
Automat==25.4.16
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
constantly==23.10.4
cryptography==45.0.5
cssselect==1.3.0
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
docstring_parser==0.16
fastapi==0.116.1
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.7.0
google-adk==1.6.1
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.103.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.35.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.2
google-crc32c==1.7.1
google-genai==1.25.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.21
grpc-google-iam-v1==0.14.2
grpcio==1.73.1
grpcio-reflection==1.73.1
grpcio-status==1.73.1
grpcio-tools==1.73.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.4
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.7.0
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-mcp-adapters==0.1.9
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langgraph==0.5.2
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.72
langgraph-swarm==0.0.12
langsmith==0.4.5
litellm==1.74.3
lxml==6.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.11.0
multidict==6.6.3
mypy_extensions==1.1.0
numpy==2.3.1
openai==1.95.1
opentelemetry-api==1.35.0
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
parsel==1.10.0
propcache==0.3.2
Protego==0.5.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyDispatcher==2.0.7
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
queuelib==1.8.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-file==2.1.0
requests-toolbelt==1.0.0
rpds-py==0.26.0
rsa==4.9.1
Scrapy==2.13.3
service-identity==24.2.0
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sse-starlette==2.4.1
starlette==0.47.1
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.9.0
tldextract==5.3.0
tokenizers==0.21.2
tqdm==4.67.1
Twisted==25.5.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
w3lib==2.3.1
watchdog==6.0.0
websockets==15.0.1
wheel==0.45.1
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zope.interface==7.2
zstandard==0.23.0
